RECIRC-29: Move grouping content in sections into component
