+++ front/main/app/styles/module/article/_content.scss
@@ -23,8 +23,8 @@
 
 	.collapsible-section-header {
 		border-bottom: 1px solid $gainsboro;
-		margin-bottom: 0.7rem;
-		padding-bottom: 0.7rem;
+		margin-bottom: .7rem;
+		padding-bottom: .7rem;
 		padding-left: 1.4rem;
 
 		.contribution-container {
